debugger: vram dump, disassembler re-enabled
